rearrange routes to prepare for login
