+++ server/package.json
@@ -11,6 +11,7 @@
   "dependencies": {
     "@headlessui/react": "^1.7.17",
     "@heroicons/react": "^2.0.18",
+    "@supabase/auth-helpers-nextjs": "^0.8.7",
     "@tailwindcss/forms": "^0.5.7",
     "axios": "^1.6.2",
     "chartjs-adapter-spacetime": "^1.0.1",
@@ -21,7 +22,8 @@
     "react": "^18",
     "react-chartjs-2": "^5.2.0",
     "react-dom": "^18",
-    "spacetime": "^7.4.8"
+    "spacetime": "^7.4.8",
+    "supabase": "^1.131.3"
   },
   "devDependencies": {
     "@types/node": "^20",
